temporary change back to singular module
