--- server/schemas/comments.js
@@ -1,7 +1,0 @@
-export const typeDef = `
-  type Comments {
-    _id: ID
-    commentText: String
-    createdAt: 
-  }
-`;
